+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add sizeLimit option to ephemeral storage (#1505)
 * Add `schedulerName` to `podTemplate` (#2114)
 * Allow overriding the auto-detected Kubernetes version
+* Garbage Collection (GC) logging disabled by default
 
 ## 0.14.0